% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{extract_product_urls}
\alias{extract_product_urls}
\title{Extract full URLs for product with a given ATC code}
\usage{
extract_product_urls(atc)
}
\arguments{
\item{atc}{string, potentially a vector}
}
\value{
character vector of full URLs to products that appear when querying pro.medicin.dk for the input ATC.
}
\description{
Extract full URLs for product with a given ATC code
}
