% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_effects.R
\name{fetch_ade_tables}
\alias{fetch_ade_tables}
\title{Fetch ADE tables from product pages}
\usage{
fetch_ade_tables(url)
}
\arguments{
\item{url}{character vector, full URL(s) for product page(s).}
}
\value{

}
\description{
ADE tables are fetches if they exist (based on a quite simplistic heuristic about their content.)
}
