% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{fetch_product_name}
\alias{fetch_product_name}
\title{Fetch product name from product page}
\usage{
fetch_product_name(url)
}
\arguments{
\item{url}{string, full URL to product page on pro.medicin.dk. Maybe be a vector of several URLs}
}
\value{
Title of product page as character.
}
\description{
Fetch product name from product page
}
