% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{find_unique_set}
\alias{find_unique_set}
\title{Find set (= unique values) from a string-list of potentially several vectors}
\usage{
find_unique_set(x, sep = ", ")
}
\arguments{
\item{x}{character, potentially a vector}

\item{sep}{character, separator. Default: \code{", "}.}
}
\value{
a scaler character.
}
\description{
Takes a character vectors whose elements are a list of values with a fixed separator, removes
duplicates, and returns a single character scalar with the same separator.
}
