% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_effects.R
\name{reconcile_ade_tables}
\alias{reconcile_ade_tables}
\title{Title}
\usage{
reconcile_ade_tables(ade_tables, ...)
}
\arguments{
\item{ade_tables}{list of data frame with appropriate layout}

\item{...}{used for recursion, anything specified will be ignored}
}
\value{

}
\description{
Title
}
