% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_effects.R
\name{simplify_ades}
\alias{simplify_ades}
\title{Make a single-line strings with all ADEs, still grouped by frequency}
\usage{
simplify_ades(ade_table, ...)
}
\arguments{
\item{ade_table}{data frame with appropriate layout}

\item{...}{
  Arguments passed on to \code{\link[=reconcile_ade_tables]{reconcile_ade_tables}}
  \describe{
    \item{\code{ade_tables}}{list of data frame with appropriate layout}
  }}
}
\value{

}
\description{
Make a single-line strings with all ADEs, still grouped by frequency
}
