% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_list}
\alias{split_list}
\title{Split in-line list containing subordinate lists inside brackets}
\usage{
split_list(x, sep = ",")
}
\arguments{
\item{x}{string containing in-line list}

\item{sep}{character, separator. Default: \code{","}.}
}
\description{
Split in-line list containing subordinate lists inside brackets
}
