% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_effects.R
\name{fetch_ade_tables}
\alias{fetch_ade_tables}
\title{Fetch ADE tables from product pages}
\usage{
fetch_ade_tables(url, sleep_time = 1)
}
\arguments{
\item{url}{character vector, full URL(s) for product page(s).}

\item{sleep_time}{number, number of seconds to hault execution, useful for avoiding loading a large number of pages in a very short amount of time.}
}
\value{

}
\description{
ADE tables are fetches if they exist (based on a quite simplistic heuristic about their content.)
}
