% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_valid}
\alias{map_valid}
\title{Map non-null list elements with function}
\usage{
map_valid(x, fun, ...)
}
\arguments{
\item{x}{list whose elements should be mapped, except for NULL elements}

\item{fun}{function to apply}

\item{...}{arguments passed to function}
}
\description{
Map non-null list elements with function
}
